+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Enhancements:
 
-- feat(compute/init): Source starter kits from service ([#1846](https://github.com/fastly/cli/pull/1846))
+- feat(compute/init): Source starter kits from service, recording `cloned_from` as a browsable link into the kit's source ([#1846](https://github.com/fastly/cli/pull/1846))
 
 ### Dependencies: